+++ tasks/html/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [1.3.1] - 2016-08-02
+
+### Modified
+- Use index file only to export the module.
+
+
 ## [1.3.0] - 2016-06-22
 
 ### Changed