+++ tasks/jspm/CHANGELOG.md
@@ -1,18 +1,27 @@
 # Change Log
 
+## [1.0.4] - 2016-08-02
+
+### Modified
+- Use index file only to export the module.
+
+
 ## [1.0.3] - 2016-05-19
 
 ### Changed
-- Fix dependencies declarations in package.json
+- Fix dependencies declarations in package.json.
+
 
 ## [1.0.2] - 2016-05-18
 
-* Update dependencies
+* Update dependencies.
+
 
 ## [1.0.1]
 
-* Update README.md
+* Update README.md.
+
 
 ## [1.0.0]
 
-* Initial release
+* Initial release.