+++ tasks/object-file/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## [1.0.1] - 2016-08-02
+
+### Modified
+- Use index file only to export the module.
+
+
 ## [1.0.0]
 
 * Initial release.